+++ src/main/java/org/socialcars/sinziana/simulation/environment/demand/jung/IInstanceJung.java
@@ -9,6 +9,6 @@ public interface IInstanceJung
 
     public String to();
 
-    public int howMany();
+    public Float howMany();
 
 }
